Refactor: Move DTO classes into dto package
