+++ src/main/java/com/spring/blackcat/magazine/MagazineService.java
@@ -1,5 +1,8 @@
 package com.spring.blackcat.magazine;
 
+import com.spring.blackcat.magazine.dto.CellDto;
+import com.spring.blackcat.magazine.dto.MagazineTitleDto;
+
 import java.util.List;
 
 public interface MagazineService {
